PBL part 3 / 18: move some C++ interpreter helpers into inline .h file.
PBL is going to use some of these same helpers, so let's put them in
`Interpreter-inl.h` rather than `Interpreter.cpp`.
